Frame/pyproject.toml
[project]
name = "frame"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ffmpeg-python>=0.2.0",
]
